Moved motor methods to CANMotor
