+++ src/CAN/CANUtils.h
@@ -56,22 +56,4 @@ deviceserial_t getDeviceSerial(const CANPacket& packet);
 
 deviceid_t getDeviceGroupAndSerial(const CANPacket& packet);
 
-// TODO: implement this and throw away the candevice stuff
-
-void initMotor(deviceserial_t serial);
-
-void initMotor(deviceserial_t serial, bool invertEncoder, bool zeroEncoder,
-			   int32_t pulsesPerJointRev, std::chrono::milliseconds telemetryPeriod);
-
-void initMotor(deviceserial_t serial, bool invertEncoder, bool zeroEncoder,
-			   int32_t pulsesPerJointRev, std::chrono::milliseconds telemetryPeriod,
-			   int32_t kP, int32_t kI, int32_t kD);
-
-void setMotorMode(deviceserial_t serial, motormode_t mode);
-
-void setMotorPower(deviceserial_t serial, double power);
-
-void setMotorPIDTarget(deviceserial_t serial, int32_t target);
-
-int32_t getMotorPosition(deviceserial_t serial);
 } // namespace can
